+++ src/collective/solr/search.py
@@ -18,7 +18,9 @@ from collective.solr.utils import getConfig
 from logging import getLogger
 from time import time
 from zope.component import queryUtility
-from zope.interface import implements
+from zope.interface import implementer
+import six
+from six.moves import map
 
 try:
     from Products.LinguaPlone.catalog import languageFilter
@@ -29,9 +31,9 @@ except ImportError:
 logger = getLogger('collective.solr.search')
 
 
+@implementer(ISearch)
 class Search(object):
     """ a search utility for solr """
-    implements(ISearch)
 
     def __init__(self):
         self.manager = None
@@ -180,7 +182,7 @@ class Search(object):
                 else:
                     query[name] = '(%s)' % ' OR '.join(value)
                 continue
-            elif isinstance(value, basestring):
+            elif isinstance(value, six.string_types):
                 if field.class_ == 'solr.TextField':
                     if isWildCard(value):
                         value = prepare_wildcard(value)